# docs/api/api-spec.yaml
swagger: "2.0"
info:
  description: "Auction Website API"
  version: "1.0.0"
  title: "Auction Website API"
host: "localhost:5000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /questions/{id}:
    get:
      tags:
      - "Q&A"
      summary: "Get question by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the question to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Question"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    
    put:
      tags:
      - "Q&A"
      summary: "Edit question by id"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: id
        type: integer
        required: true
        description: Numeric ID of the quesiton to edit.
      - in: "body"
        name: "body"
        description: "Created question object"
        required: true
        schema:
          $ref: "#/definitions/QuestionReq"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Question"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
    delete:
      tags:
      - "Q&A"
      summary: "Delete question by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the question to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Question"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /questions:
    post:
      tags:
      - "Q&A"
      summary: "Post new question"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created question object"
        required: true
        schema:
          $ref: "#/definitions/QuestionReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Question"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
      
    get:
      tags:
      - "Q&A"
      summary: "Get questions with answers"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "page"
        description: "Page number"
        required: true
        type: integer
      - in: "query"
        name: "auctionId"
        description: "auction Id"
        required: false
        type: integer
      - in: "query"
        name: "askerId"
        description: "asker Id"
        required: false
        type: integer
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/QuestionsList"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
            
    delete:
      tags:
      - "Q&A"
      summary: "Delete ALL questions"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GenericResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"          
  /answers/{id}:
    get:
      tags:
      - "Q&A"
      summary: "Get answer by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the answer to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Answer"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    put:
      tags:
      - "Q&A"
      summary: "Edit answer by id"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: id
        type: integer
        required: true
        description: Numeric ID of the answer to edit.
      - in: "body"
        name: "body"
        description: "Created answer object"
        required: true
        schema:
          $ref: "#/definitions/AnswerReq"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Answer"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
      
    delete:
      tags:
      - "Q&A"
      summary: "Delete answer by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the answer to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Answer"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /answers:
    post:
      tags:
      - "Q&A"
      summary: "Post new answer"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created answer"
        required: true
        schema:
          $ref: "#/definitions/AnswerReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Question"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
  /item_meta/subcategories/{id}:
    get:
      tags:
      - "item_meta"
      summary: "Get subcategory by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the subcategory to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubCategory"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    put:
      tags:
      - "item_meta"
      summary: "Get subcategory by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the subcategory to edit.
        - in: "body"
          name: "body"
          description: "Created subcategory object"
          required: true
          schema:
            $ref: "#/definitions/SubCategoryReq"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubCategory"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    
    delete:
      tags:
      - "item_meta"
      summary: "Delete subcategory by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the subcategory to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubCategory"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /item_meta/subcategories:
    post:
      tags:
      - "item_meta"
      summary: "Create new subcategory"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created subcategory"
        required: true
        schema:
          $ref: "#/definitions/SubCategoryReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubCategory"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    get:
      tags:
      - "item_meta"
      summary: "Get subcategories"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filters for get subcategories"
        required: true
        schema:
          $ref: "#/definitions/GetSubCategories"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/SubCategoriesList"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
            
    delete:
      tags:
      - "item_meta"
      summary: "Delete ALL subcategories"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GenericResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
  /items/{id}:
    get:
      tags:
      - "item"
      summary: "Get item by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the item to get.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Item"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    put:
      tags:
      - "item"
      summary: "Edit item by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the item to edit.
        - in: "body"
          name: "body"
          description: "Created item object"
          required: true
          schema:
            $ref: "#/definitions/ItemReq"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Item"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    
    delete:
      tags:
      - "item"
      summary: "Delete item by id"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the item to delete.
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Item"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /items:
    post:
      tags:
      - "item"
      summary: "Create new item"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created item"
        required: true
        schema:
          $ref: "#/definitions/ItemReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Item"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    get:
      tags:
      - "item"
      summary: "Get items"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "page"
        description: "Page number"
        required: true
        type: integer
      - in: "query"
        name: "categoryId"
        description: "CategoryId"
        required: false
        type: integer
      - in: "query"
        name: "subcategoryId"
        description: "SubCategoryId"
        required: false
        type: integer
      - in: "query"
        name: "attributeId"
        description: "AttributeId"
        required: false
        type: integer
      - in: "query"
        name: "attributeValue"
        description: "AttributeValue"
        required: false
        type: string
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ItemsList"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
            
            
    delete:
      tags:
      - "item"
      summary: "Delete ALL items"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GenericResponse"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  QuestionReq:
    type: "object"
    properties:
      auctionId:
        type: "integer"
        format: "int32"
      questionText:
        type: "string"
  Question:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      auctionId:
        type: "integer"
        format: "int32"
      questionText:
        type: "string"
      replyText:
        type: "string"
      isOpen:
        type: "boolean"
      askerId:
        type: "integer"
        format: "int32"
      replierId:
        type: "integer"
        format: "int32"
...
  QuestionsList:
    type: "array"
    items:
      $ref: "#/definitions/Question"
  AnswerReq:
    type: "object"
    properties:
      questionId:
        type: "integer"
        format: "int32"
      replyText:
        type: "string"
  Answer:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      questionId:
        type: "integer"
        format: "int32"
      replyText:
        type: "string"
      replierId:
        type: "integer"
        format: "int632"
...
  SubCategoryReq:
    type: "object"
    properties:
      categoryId:
        type: "integer"
        format: "int32"
      subcategoryName:
        type: "string"
  SubCategory:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      categoryId:
        type: "integer"
        format: "int32"
      subcategoryName:
        type: "string"
  SubCategoriesList:
    type: "array"
    items:
      $ref: "#/definitions/SubCategory"
  GetSubCategories:
    type: "object"
    properties:
      offset:
        type: "integer"
        format: "int32"
      count:
        type: "integer"
        format: "int32"
      sort:
        type: "string"
...
  ItemAttributeReq:
    type: "object"
    properties:
      attributeId:
        type: "integer"
        format: "int32"
      attributeValue:
        type: "string"
  ItemAttributeReqsList:
    type: "array"
    items:
      $ref: "#/definitions/ItemAttributeReq"
  ItemAttribute:
    type: "object"
    properties:
      attributeId:
        type: "integer"
        format: "int32"
      attributeName:
        type: "string"
      attributeValue:
        type: "string"
  ItemAttributesList:
    type: "array"
    items:
      $ref: "#/definitions/ItemAttribute"
  ItemReq:
    type: "object"
    properties:
      categoryId:
        type: "integer"
        format: "int32"
      subcategoryId:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      attributes:
        $ref: "#/definitions/ItemAttributeReqsList"
  Item:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      categoryId:
        type: "integer"
        format: "int32"
      categoryName:
        type: "string"
      subcategoryId:
        type: "integer"
        format: "int32"
      subcategoryName:
        type: "string"
      name:
        type: "string"
      attributes:
        $ref: "#/definitions/ItemAttributesList"
  ItemsList:
    type: "array"
    items:
      $ref: "#/definitions/Item"
...
  GenericResponse:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - code
      - message
  # Schema for error response body
  Error:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - code
      - message